Move some notification data related structs into a separate file.
