+++ com.unity.mobile.notifications/Runtime/iOS/Plugins/UnityAppController+Notifications.h
@@ -1,15 +1,13 @@
 //
 //  UnityAppController+Notifications.h
-//  Unity-iPhone
+//  iOS.notifications
 //
-//  Copyright © 2018 Unity Technologies. All rights reserved.
-//
+
 #if TARGET_OS_IOS
 #import "UnityAppController.h"
 
 #include "Classes/PluginBase/LifeCycleListener.h"
 #include "Classes/PluginBase/AppDelegateListener.h"
-
 
 @interface UnityAppController (Notifications)
