+++ com.unity.mobile.notifications/Runtime/iOS/Plugins/UnityNotificationData.h.meta
@@ -1,0 +1,26 @@
+fileFormatVersion: 2
+guid: 7609257501e8b48f98214be5de026339
+PluginImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  iconMap: {}
+  executionOrder: {}
+  isPreloaded: 0
+  isOverridable: 1
+  isExplicitlyReferenced: 0
+  validateReferences: 1
+  platformData:
+  - first:
+      Any: 
+    second:
+      enabled: 1
+      settings: {}
+  - first:
+      Editor: Editor
+    second:
+      enabled: 0
+      settings:
+        DefaultValueInitialized: true
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
