+++ com.unity.mobile.notifications/Runtime/iOS/Plugins/UnityNotificationManager.h
@@ -2,88 +2,15 @@
 //  UnityNotificationManager.h
 //  iOS.notifications
 //
-//  Copyright © 2018 Unity Technologies. All rights reserved.
-//
 
 #if TARGET_OS_IOS
 
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
 #import <UserNotifications/UserNotifications.h>
-//
+#import "UnityNotificationData.h"
+
 #define SYSTEM_VERSION_10_OR_ABOVE  ([[[UIDevice currentDevice] systemVersion] compare:@"10.0" options:NSNumericSearch] != NSOrderedAscending)
-
-enum triggerType
-{
-    TIME_TRIGGER = 0,
-    CALENDAR_TRIGGER = 10,
-    LOCATION_TRIGGER = 20,
-    PUSH_TRIGGER = 3,
-};
-
-typedef struct iOSNotificationData
-{
-    char* identifier;
-    char* title;
-    char* body;
-    int badge;
-    char* subtitle;
-    char* categoryIdentifier;
-    char* threadIdentifier;
-
-    //Custom data
-    char* data;
-    BOOL showInForeground;
-    int showInForegroundPresentationOptions;
-
-    // Trigger
-    int triggerType;  //0 - time, 1 - calendar, 2 - location, 3 - push.
-    BOOL repeats;
-
-    //Time trigger
-    int timeTriggerInterval;
-
-    //Calendar trigger
-    int calendarTriggerYear;
-    int calendarTriggerMonth;
-    int calendarTriggerDay;
-    int calendarTriggerHour;
-    int calendarTriggerMinute;
-    int calendarTriggerSecond;
-
-    //Location trigger
-    float locationTriggerCenterX;
-    float locationTriggerCenterY;
-    float locationTriggerRadius;
-    bool locationTriggerNotifyOnEntry;
-    bool locationTriggerNotifyOnExit;
-} iOSNotificationData;
-
-
-typedef struct iOSNotificationAuthorizationData
-{
-    bool granted;
-    char* error;
-    bool finished;
-    const char* deviceToken;
-} iOSNotificationAuthorizationData;
-
-
-typedef void (*NotificationDataReceivedResponse)(struct iOSNotificationData* data);
-typedef void (*AuthorizationRequestResponse) (struct iOSNotificationAuthorizationData* data);
-
-typedef struct NotificationSettingsData
-{
-    int authorizationStatus;
-    int notificationCenterSetting;
-    int lockScreenSetting;
-    int carPlaySetting;
-    int alertSetting;
-    int badgeSetting;
-    int soundSetting;
-    int alertStyle;
-    int showPreviewsSetting;
-} NotificationSettingsData;
 
 @interface UnityNotificationManager : NSObject<UNUserNotificationCenterDelegate>
