+++ com.unity.mobile.notifications/Runtime/iOS/Plugins/UnityNotificationManager.m
@@ -2,15 +2,13 @@
 //  UnityNotificationManager.m
 //  iOS.notifications
 //
-//  Created by Paulius on 24/07/2018.
-//  Copyright © 2018 Unity Technologies. All rights reserved.
-//
+
 #if TARGET_OS_IOS
 
 #import "UnityNotificationManager.h"
 
 #if defined(UNITY_USES_LOCATION) && UNITY_USES_LOCATION
-    #import <CoreLocation/CoreLocation.h>
+#import <CoreLocation/CoreLocation.h>
 #endif
 
 const int kDefaultPresentationOptions = -1;
